pyproject.toml
[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[project]
name = "pygrind"
version = "1.0.0"
description = "PyGrind — Python Competition Grind"
requires-python = ">=3.10"
license = "MIT"
dependencies = [
    "PyQt6>=6.5",
    "QScintilla>=2.14",
    "PyYAML>=6.0",
    "platformdirs>=4.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-qt>=4.2",
    "ruff>=0.4",
    "pyinstaller>=6.0",
]

[project.scripts]
pygrind = "pygrind.app:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 99
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "B", "SIM"]

[tool.ruff.format]
quote-style = "double"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
